scripts/requirements.txt
requests>=2.28.0
openai>=0.27.0
google-cloud-storage>=2.7.0
google-auth>=2.16.0
google-api-python-client>=2.70.0
feedparser>=6.0.10
praw>=7.6.0
tweepy>=4.12.0